Move MenuItem class definition to its own file
We're going to be adding more configurable options to this area and we
should not just continue to define classes in this file.
